# docs/api-spec.yaml
openapi: 3.0.3
info:
  title: Restaurant POS System API
  description: |
    RESTful API for the Restaurant Point of Sale System.
    
    ## Authentication
    All endpoints require authentication unless otherwise noted. The API supports:
    - **Session-based auth** (web clients): Uses Flask-Login sessions
    - **JWT Bearer tokens** (API clients): Include `Authorization: Bearer <token>` header
    - **API Keys** (integrations): Include `X-API-Key: <key>` header
    
    ## Rate Limiting
    - Standard endpoints: 100 requests/minute
    - Search endpoints: 30 requests/minute
    - Batch operations: 10 requests/minute
    
    ## Pagination
    List endpoints support pagination via query parameters:
    - `page` (default: 1)
    - `per_page` (default: 20, max: 100)
    
    Response includes pagination metadata in `meta` object.
    
    ## Error Responses
    All errors follow a consistent format with `error`, `message`, and optional `details`.
  version: 1.0.0
  contact:
    name: API Support
    email: api-support@restaurant-pos.com
  license:
    name: Proprietary
servers:
  - url: https://api.restaurant-pos.com/api/v1
    description: Production
  - url: https://staging-api.restaurant-pos.com/api/v1
    description: Staging
  - url: http://localhost:5000/api/v1
    description: Local Development
...
paths:
...
  /menu/categories:
    get:
      tags: [Menu]
      summary: List all categories
      operationId: listCategories
      parameters:
        - $ref: '#/components/parameters/includeInactive'
        - name: parent_id
          in: query
          schema:
            type: integer
          description: Filter by parent category (null for top-level)
      responses:
        '200':
          description: List of categories
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Category'
    post:
      tags: [Menu]
      summary: Create category
      operationId: createCategory
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CategoryCreate'
      responses:
        '201':
          description: Category created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Category'
        '400':
          $ref: '#/components/responses/BadRequest'
        '403':
          $ref: '#/components/responses/Forbidden'
...
  /menu/items:
    get:
      tags: [Menu]
      summary: List menu items
      operationId: listMenuItems
      parameters:
        - $ref: '#/components/parameters/page'
        - $ref: '#/components/parameters/perPage'
        - $ref: '#/components/parameters/includeInactive'
        - name: category_id
          in: query
          schema:
            type: integer
        - name: search
          in: query
          schema:
            type: string
          description: Search by name or SKU
        - name: available_only
          in: query
          schema:
            type: boolean
            default: false
      responses:
        '200':
          description: Paginated list of menu items
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/MenuItem'
                  meta:
                    $ref: '#/components/schemas/PaginationMeta'
    post:
      tags: [Menu]
      summary: Create menu item
      operationId: createMenuItem
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MenuItemCreate'
      responses:
        '201':
          description: Menu item created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MenuItem'
  /menu/items/{id}:
    parameters:
      - $ref: '#/components/parameters/resourceId'
    get:
      tags: [Menu]
      summary: Get menu item by ID
      operationId: getMenuItem
      parameters:
        - name: include
          in: query
          schema:
            type: string
            enum: [modifiers, pricing_rules, recipe]
          description: Include related data
      responses:
        '200':
          description: Menu item details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MenuItemFull'
    put:
      tags: [Menu]
      summary: Update menu item
      operationId: updateMenuItem
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MenuItemUpdate'
      responses:
        '200':
          description: Menu item updated
    patch:
      tags: [Menu]
      summary: Partial update menu item
      operationId: patchMenuItem
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                is_available:
                  type: boolean
                price:
                  type: number
      responses:
        '200':
          description: Menu item patched
    delete:
      tags: [Menu]
      summary: Delete menu item
      operationId: deleteMenuItem
      responses:
        '204':
          description: Menu item deleted
...
  /menu/modifier-groups:
    get:
      tags: [Menu]
      summary: List modifier groups
      operationId: listModifierGroups
      responses:
        '200':
          description: List of modifier groups
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/ModifierGroup'
    post:
      tags: [Menu]
      summary: Create modifier group
      operationId: createModifierGroup
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ModifierGroupCreate'
      responses:
        '201':
          description: Modifier group created
  /menu/modifiers:
    get:
      tags: [Menu]
      summary: List modifiers
      operationId: listModifiers
      parameters:
        - name: group_id
          in: query
          schema:
            type: integer
      responses:
        '200':
          description: List of modifiers
    post:
      tags: [Menu]
      summary: Create modifier
      operationId: createModifier
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ModifierCreate'
      responses:
        '201':
          description: Modifier created

  # ==================== ORDERS ====================
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    apiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
    sessionAuth:
      type: apiKey
      in: cookie
      name: session
  parameters:
    resourceId:
      name: id
      in: path
      required: true
      schema:
        type: integer
      description: Resource ID
    page:
      name: page
      in: query
      schema:
        type: integer
        default: 1
        minimum: 1
    perPage:
      name: per_page
      in: query
      schema:
        type: integer
        default: 20
        minimum: 1
        maximum: 100
    includeInactive:
      name: include_inactive
      in: query
      schema:
        type: boolean
        default: false
  responses:
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: "bad_request"
            message: "Invalid request parameters"
            details:
              field: "price"
              issue: "must be a positive number"
...
    Forbidden:
      description: Permission denied
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: "forbidden"
            message: "You don't have permission to perform this action"
...
  schemas:
    Error:
      type: object
      required: [error, message]
      properties:
        error:
          type: string
        message:
          type: string
        details:
          type: object
    PaginationMeta:
      type: object
      properties:
        page:
          type: integer
        per_page:
          type: integer
        total:
          type: integer
        total_pages:
          type: integer

    # Auth Schemas
...
    Category:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
        display_order:
          type: integer
        parent_id:
          type: integer
        is_active:
          type: boolean
        image_url:
          type: string
        item_count:
          type: integer
    CategoryCreate:
      type: object
      required: [name]
      properties:
        name:
          type: string
        description:
          type: string
        display_order:
          type: integer
        parent_id:
          type: integer
        image_url:
          type: string
...
    MenuItem:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
        price:
          type: number
        category_id:
          type: integer
        category_name:
          type: string
        sku:
          type: string
        image_url:
          type: string
        is_available:
          type: boolean
        is_taxable:
          type: boolean
        prep_time_minutes:
          type: integer
        display_order:
          type: integer
    MenuItemFull:
      allOf:
        - $ref: '#/components/schemas/MenuItem'
        - type: object
          properties:
            modifier_groups:
              type: array
              items:
                $ref: '#/components/schemas/ModifierGroup'
            pricing_rules:
              type: array
              items:
                $ref: '#/components/schemas/PricingRule'
    MenuItemCreate:
      type: object
      required: [name, price, category_id]
      properties:
        name:
          type: string
        description:
          type: string
        price:
          type: number
        category_id:
          type: integer
        sku:
          type: string
        image_url:
          type: string
        is_taxable:
          type: boolean
          default: true
        prep_time_minutes:
          type: integer
        modifier_group_ids:
          type: array
          items:
            type: integer
    MenuItemUpdate:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        price:
          type: number
        category_id:
          type: integer
        sku:
          type: string
        image_url:
          type: string
        is_available:
          type: boolean
        is_taxable:
          type: boolean
        prep_time_minutes:
          type: integer
        display_order:
          type: integer
    ModifierGroup:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        min_selections:
          type: integer
        max_selections:
          type: integer
        is_required:
          type: boolean
        display_order:
          type: integer
        modifiers:
          type: array
          items:
            $ref: '#/components/schemas/Modifier'
    ModifierGroupCreate:
      type: object
      required: [name]
      properties:
        name:
          type: string
        min_selections:
          type: integer
          default: 0
        max_selections:
          type: integer
          default: 1
        is_required:
          type: boolean
          default: false
    Modifier:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        price:
          type: number
        modifier_group_id:
          type: integer
        is_available:
          type: boolean
        display_order:
          type: integer
    ModifierCreate:
      type: object
      required: [name, modifier_group_id]
      properties:
        name:
          type: string
        price:
          type: number
          default: 0
        modifier_group_id:
          type: integer
    PricingRule:
      type: object
      properties:
        id:
          type: integer
        menu_item_id:
          type: integer
        rule_type:
          type: string
        price:
          type: number
        start_time:
          type: string
          format: time
        end_time:
          type: string
          format: time
        days_of_week:
          type: array
          items:
            type: integer
        priority:
          type: integer

    # Order Schemas
...
security:
  - bearerAuth: []
  - sessionAuth: []
  - apiKeyAuth: []